Create a new directory include/modmesh/mesh/ for unstructured-mesh code
Use std::algorithm to count maximal number of faces for interior data.
