+++ .clang-tidy
@@ -23,6 +23,7 @@ Checks: >
     -performance-noexcept-move-constructor,
     -readability-named-parameter,
     -readability-magic-numbers,
+    -readability-redundant-access-specifiers,
     -misc-non-private-member-variables-in-classes
 WarningsAsErrors: ''
 HeaderFilterRegex: 'modmesh.*'
